+++ src/app/CompendiumMatcher.js
@@ -44,14 +44,15 @@ export class CompendiumMatcher {
   getNameMatch(pbName, foundryName) {
     for (const [packName, index] of Object.entries(this.indexes)) {
       logger.debug(`Checking for compendium documents for ${pbName} (${foundryName}) in ${packName}`);
-      const indexMatch = index.find((i) => i.system.slug === Seasoning.slug(foundryName))
-        ?? index.find((i) => i.system.slug === Seasoning.slug(pbName));
+      const indexMatch = index.find((i) => (i.system.slug ?? Seasoning.slug(i.name)) === Seasoning.slug(foundryName))
+        ?? index.find((i) => (i.system.slug ?? Seasoning.slug(i.name)) === Seasoning.slug(pbName));
 
       if (indexMatch) {
         logger.debug(`Found compendium document for ${pbName} (${foundryName}) in ${packName} with id ${indexMatch._id}`);
         return { i: indexMatch, pack: this.packs[packName] };
       }
     }
+
     return undefined;
   }